moved test helper file
