+++ src/app/features/completion/components/playing-item/playing-item.component.spec.ts
@@ -18,7 +18,7 @@ import * as fromCompletion from '../../reducers/root.reducer';
 
 import { PlayingDisplayData } from '../../models';
 
-import { user } from '../../../../testing-helpers';
+import { user } from '../../../../testing/testing-helpers';
 
 describe('PlayingItemComponent', () => {
   let store: Store<fromRoot.State>;
